Add deleted IffFile and switch to new namespace
