--- PangLib.IFF/Model/Data/Enchant.cs
@@ -1,11 +1,0 @@
-using System.Runtime.InteropServices;
-
-namespace PangLib.IFF.Model.Data;
-
-[StructLayout(LayoutKind.Sequential, Pack = 4)]
-public struct Enchant
-{
-    public uint Active { get; set; }
-    public uint ID { get; set; }
-    public uint Price { get; set; }
-}
