--- PangLib.IFF/Model/Data/Mascot.cs
@@ -1,19 +1,0 @@
-using System.Runtime.InteropServices;
-using PangLib.IFF.Model.General;
-
-namespace PangLib.IFF.Model.Data;
-
-[StructLayout(LayoutKind.Sequential, Pack = 4)]
-public struct Mascot
-{
-    [field: MarshalAs(UnmanagedType.Struct)]
-    public IFFCommonUs8 Header { get; set; }
-    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
-    public string Texture1 { get; set; }
-    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
-    public string Texture2 { get; set; }
-    public ushort Price1Day { get; set; }
-    public ushort Price7Day { get; set; }
-    public ushort Unknown1 { get; set; }
-    public ushort Price30Day { get; set; }
-}
