--- PangLib.IFF/Model/Type/ClubSetWorkshopType.cs
@@ -1,7 +1,0 @@
-﻿namespace PangLib.IFF.Model.Type;
-
-public enum ClubSetWorkshopType : int
-{
-    NotUpgradable = -1,
-    Upgrdable = 0
-}
