--- PangLib.IFF/Model/Type/PartType.cs
@@ -1,14 +1,0 @@
-﻿namespace PangLib.IFF.Model.Type;
-
-public enum PartType : uint
-{
-    Top = 0,
-    Bottom = 1,
-    Hat = 2,
-    Gloves = 3,
-    Shoes = 4,
-    AccesoryOrBase = 5,
-    SubLeg = 6,
-    UccBlank = 8,
-    UccCopy = 9
-}
