--- PangLib.IFF/Model/Type/TimeLimitedType.cs
@@ -1,7 +1,0 @@
-﻿namespace PangLib.IFF.Model.Type;
-
-public enum TimeLimitedType : uint
-{
-    NotTimeLimited = 0,
-    TimeLimited = 1
-}
