+++ PangLib.IFF/Model/US/Season8/Data/Ball.cs
@@ -1,0 +1,53 @@
+using System.Runtime.InteropServices;
+using PangLib.IFF.Model.US.Season8.General;
+
+namespace PangLib.IFF.Model.US.Season8.Data;
+
+[StructLayout(LayoutKind.Sequential, Pack = 4)]
+public struct Ball
+{
+    [field: MarshalAs(UnmanagedType.Struct)]
+    public IffCommon Header { get; set; }
+    
+    public uint Unknown { get; set; }
+    
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string Model { get; set; }
+    
+    public uint Unknown2 { get; set; }
+    public uint Unknown3 { get; set; }
+    
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence1 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence2 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence3 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence4 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence5 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence6 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallSequence7 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx1 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx2 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx3 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx4 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx5 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx6 { get; set; }
+    [field: MarshalAs(UnmanagedType.ByValTStr, SizeConst = 40)]
+    public string BallFx7 { get; set; }
+    
+    [field: MarshalAs(UnmanagedType.Struct)]
+    public StatsStruct Stats { get; set; }
+    
+    public ushort Unknown4 { get; set; }
+}
