+++ PangLib.IFF/Model/US/Season8/Type/ClubSetWorkshopType.cs
@@ -1,0 +1,7 @@
+﻿namespace PangLib.IFF.Model.US.Season8.Type;
+
+public enum ClubSetWorkshopType : int
+{
+    NotUpgradable = -1,
+    Upgrdable = 0
+}
