+++ PangLib.IFF/Model/US/Season8/Type/PartType.cs
@@ -1,0 +1,14 @@
+﻿namespace PangLib.IFF.Model.US.Season8.Type;
+
+public enum PartType : uint
+{
+    Top = 0,
+    Bottom = 1,
+    Hat = 2,
+    Gloves = 3,
+    Shoes = 4,
+    AccesoryOrBase = 5,
+    SubLeg = 6,
+    UccBlank = 8,
+    UccCopy = 9
+}
