+++ PangLib.IFF/Model/US/Season8/Type/ShopTempItemDurationType.cs
@@ -1,0 +1,11 @@
+﻿namespace PangLib.IFF.Model.US.Season8.Type;
+
+public enum ShopTempItemDurationType : byte
+{
+    None = 0,
+    Day = 1,
+    Week = 7,
+    TwoWeeks = 15,
+    Month = 30,
+    Year = 0x6D
+}
