+++ PangLib.IFF/Model/US/Season8/Type/TimeLimitedType.cs
@@ -1,0 +1,7 @@
+﻿namespace PangLib.IFF.Model.US.Season8.Type;
+
+public enum TimeLimitedType : uint
+{
+    NotTimeLimited = 0,
+    TimeLimited = 1
+}
